Add 02:05 (GMT) to schedule
After daylight saving time happened (or stopped, whatever), the badge of the day id is no longer updated at previous time. I believe that the action runs at the correct time, so this may be the Kongregate homepage that is updated an hour late.
Whatever the issue is, the workflow will run an additional time every day to work around it.